добавлены тесты репозитория по формулам, запись и чтение

+++ repository/src/main/java/com/pet/animal/formula/dose/health/veterinary/cure/repo/dao/VetFormulaDao.kt
@@ -26,7 +26,7 @@ interface VetFormulaDao {
     suspend fun deleteFormula(formulaEntity: FormulaEntity)
 
     @Query("delete from defFormulas where id=:id")
-    suspend fun deleteFormulaByID(id:Int)
+    suspend fun deleteFormulaByID(id:Long)
 
     //URLS
